+++ cccm-frontend/src/App.vue
@@ -24,6 +24,7 @@ export default {
   },
   data() {
     return {
+        // default location
         locationVal: "Victoria Probation Office"
     }
   },
